chore(packages/anzii): set validation parameter
Absolute paths without base fail when a check is made for the file path
existence validation.

The failing validation is due to nodejs's resolution mechanism, which
appends a file protoco scheme to the beginning of the path, since this
path doesn't have base, the existence of the path fails.

Setting a checkExist parameter will ensure that a user decides if we
should bypass validation, for example to let nodejs resolves, perhaps with
a custom resolver.