chore(packages/anzii): adds support for static absolute root
This change enables anzii to use an absolute path as provided by the
user for serving static files. It requires that the domain config entry that
the users passes contains 'absolute' key.

The 'dumain.absolute` property instructs anziijs to skip creating the
absolute path by combining the document root and value passed by the user
in 'dumain.set'