chore(packages/anzii): creates logs config object
The new log object created in this commit will be used to help improve the
performance of logging in anziijs.

As opposed to the previous implementation where we had fire an event to
anziiloger plugin to determine logging, we now do it at a base level where
we can quickly return if a condition disables logging of certain logs.